requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.0.0
sqlparse==0.5.4
tzdata==2025.2
